+++ titiler/cmr/backend.py
@@ -2,8 +2,20 @@
 
 import os
 import re
+from collections.abc import Callable, Mapping
 from datetime import datetime
-from typing import Any, Dict, List, Literal, Optional, Tuple, Type, TypedDict, Union
+from typing import (
+    Any,
+    Dict,
+    List,
+    Literal,
+    Optional,
+    Protocol,
+    Tuple,
+    Type,
+    TypedDict,
+    Union,
+)
 
 import attr
 import earthaccess
@@ -11,18 +23,20 @@ import rasterio
 import rasterio.session
 from cachetools import TTLCache, cached
 from cachetools.keys import hashkey
-from cogeo_mosaic.backends import BaseBackend
+from cogeo_mosaic.backends.base import BaseBackend
 from cogeo_mosaic.errors import NoAssetFoundError
 from cogeo_mosaic.mosaic import MosaicJSON
-from earthaccess.auth import Auth
-from morecantile import Tile, TileMatrixSet
-from rasterio.crs import CRS
+from earthaccess.results import DataGranule
+from morecantile.commons import Tile
+from morecantile.models import TileMatrixSet
+from rasterio import CRS
 from rasterio.features import bounds
 from rasterio.warp import transform_bounds, transform_geom
 from rio_tiler.constants import WEB_MERCATOR_TMS, WGS84_CRS
-from rio_tiler.io import BaseReader, Reader
+from rio_tiler.io.base import BaseReader
+from rio_tiler.io.rasterio import Reader
 from rio_tiler.models import ImageData
-from rio_tiler.mosaic import mosaic_reader
+from rio_tiler.mosaic.reader import mosaic_reader
 from rio_tiler.types import BBox
 
 from titiler.cmr.logger import logger
@@ -36,20 +50,30 @@ retry_config = RetrySettings()
 s3_auth_config = AuthSettings()
 
 
-@cached(  # type: ignore
-    TTLCache(maxsize=100, ttl=60),
-    key=lambda auth, daac: hashkey(auth.token["access_token"], daac),
-)
-def aws_s3_credential(auth: Auth, provider: str) -> Dict:
-    """Get AWS S3 credential through earthaccess."""
-    return auth.get_s3_credentials(provider=provider)
+class AWSCredentials(TypedDict, total=True):
+    """AWS S3 temporary credentials."""
+
+    accessKeyId: str
+    secretAccessKey: str
+    sessionToken: str
+    # Parseable with datetime.fromisoformat (e.g., 2025-12-12 21:07:02+00:00)
+    expiration: str
 
 
-class Asset(TypedDict, total=False):
+class Asset(TypedDict, total=True):
     """Simple Asset model."""
 
-    url: Union[str, Dict[str, str]]
+    url: Union[str, Mapping[str, str]]
     provider: str
+    s3_credentials_url: str | None
+
+
+class FindGranules(Protocol):
+    """Protocol for functions that search for granules in CMR."""
+
+    def __call__(self, count: int, **kwargs: Any) -> list[DataGranule]:
+        """Match signature of earthaccess.search_data."""
+        ...
 
 
 @attr.s
@@ -72,11 +96,14 @@ class CMRBackend(BaseBackend):
     # The reader is read-only (outside init)
     mosaic_def: MosaicJSON = attr.ib(init=False)
 
-    auth: Optional[Auth] = attr.ib(default=None)
-
     input: str = attr.ib("CMR", init=False)
 
     _backend_name: str = attr.ib(default="CMR")
+
+    find_granules: FindGranules = attr.ib(default=earthaccess.search_data)
+    auth: earthaccess.Auth | None = attr.ib(default=None)
+    get_s3_credentials: Callable[[str], AWSCredentials] | None = attr.ib(default=None)
+    rasterio_env_kwargs: Mapping[str, Any] = attr.ib(factory=dict)
 
     def __attrs_post_init__(self) -> None:
         """Post Init."""
@@ -112,35 +139,37 @@ class CMRBackend(BaseBackend):
         """This method is not used but is required by the abstract class."""
         pass
 
-    def _get_s3_credentials(self, asset: Asset) -> Optional[Dict]:
+    def _get_s3_credentials(self, asset: Asset) -> Optional[AWSCredentials]:
         """Get s3 credentials from kwargs or via auth."""
-        if (
-            s3_auth_config.strategy == "environment"
-            and s3_auth_config.access == "direct"
-            and self.auth
-        ):
-            return aws_s3_credential(self.auth, asset["provider"])
-        else:
-            return None
+        return (
+            self.get_s3_credentials(endpoint)
+            if self.get_s3_credentials
+            and (endpoint := asset.get("s3_credentials_url", None))
+            else None
+        )
 
-    def _build_reader_options(self, s3_credentials: Optional[Dict]) -> Dict:
+    def _build_reader_options(self, s3_credentials: Optional[AWSCredentials]) -> Dict:
         """Build reader options with opener_options if s3_credentials provided."""
-        if s3_credentials:
-            return {
-                **self.reader_options,
-                "opener_options": {
-                    "s3_credentials": {
-                        "key": s3_credentials["accessKeyId"],
-                        "secret": s3_credentials["secretAccessKey"],
-                        "token": s3_credentials["sessionToken"],
-                    }
+        return {
+            **self.reader_options,
+            "opener_options": {
+                "auth": self.auth,
+                **{
+                    "s3_credentials": (
+                        {
+                            "key": s3_credentials["accessKeyId"],
+                            "secret": s3_credentials["secretAccessKey"],
+                            "token": s3_credentials["sessionToken"],
+                        }
+                        if s3_credentials
+                        else {}
+                    )
                 },
-            }
-        else:
-            return self.reader_options
+            },
+        }
 
     def _create_aws_session(
-        self, s3_credentials: Optional[Dict]
+        self, s3_credentials: Optional[AWSCredentials]
     ) -> Optional[rasterio.session.AWSSession]:
         """Create rasterio AWSSession from s3 credentials."""
         if s3_credentials:
@@ -152,9 +181,10 @@ class CMRBackend(BaseBackend):
         return None
 
     def assets_for_tile(
-        self, x: int, y: int, z: int, access: Access = "direct", **kwargs: Any
+        self, x: int, y: int, z: int, access: Access | None = None, **kwargs: Any
     ) -> List[Asset]:
         """Retrieve assets for tile."""
+        access = access or s3_auth_config.access
         bbox = self.tms.bounds(Tile(x, y, z))
         return self.get_assets(*bbox, access=access, **kwargs)
 
@@ -175,10 +205,12 @@ class CMRBackend(BaseBackend):
         xmax: float,
         ymax: float,
         coord_crs: CRS = WGS84_CRS,
-        access: Access = "direct",
+        access: Access | None = None,
         **kwargs: Any,
     ) -> List[Asset]:
         """Retrieve assets for bbox."""
+        access = access or s3_auth_config.access
+
         if coord_crs != WGS84_CRS:
             xmin, ymin, xmax, ymax = transform_bounds(
                 coord_crs,
@@ -210,10 +242,11 @@ class CMRBackend(BaseBackend):
         ymax: float,
         limit: int = 100,
         bands_regex: Optional[str] = None,
-        access: Access = "direct",
+        access: Access | None = None,
         **kwargs: Any,
     ) -> List[Asset]:
         """Find assets."""
+        access = access or s3_auth_config.access
         xmin, ymin, xmax, ymax = (round(n, 8) for n in [xmin, ymin, xmax, ymax])
         assets: List[Asset] = []
 
@@ -224,12 +257,12 @@ class CMRBackend(BaseBackend):
             if isinstance(temporal, datetime):
                 kwargs["temporal"] = (temporal, temporal)
         try:
-            results = earthaccess.search_data(
+            results = self.find_granules(
                 bounding_box=(xmin, ymin, xmax, ymax),
                 count=limit,
                 **kwargs,
             )
-        except IndexError:
+        except Exception:
             return assets
 
         for r in results:
@@ -247,6 +280,7 @@ class CMRBackend(BaseBackend):
                         {
                             "url": urls,
                             "provider": r["meta"]["provider-id"],
+                            "s3_credentials_url": r.get_s3_credentials_endpoint(),
                         }
                     )
 
@@ -255,6 +289,7 @@ class CMRBackend(BaseBackend):
                     {
                         "url": r.data_links(access=access)[0],
                         "provider": r["meta"]["provider-id"],
+                        "s3_credentials_url": r.get_s3_credentials_endpoint(),
                     }
                 )
 
@@ -271,16 +306,18 @@ class CMRBackend(BaseBackend):
         tile_z: int,
         cmr_query: Dict,
         bands_regex: Optional[str] = None,
+        access: Access | None = None,
         **kwargs: Any,
     ) -> Tuple[ImageData, List[str]]:
         """Get Tile from multiple observation."""
         logger.info("searching for assets")
+        access = access or s3_auth_config.access
         mosaic_assets = self.assets_for_tile(
             tile_x,
             tile_y,
             tile_z,
             **cmr_query,
-            access=s3_auth_config.access,
+            access=access,
             bands_regex=bands_regex,
         )
         logger.info(f"found {len(mosaic_assets)} assets")
@@ -293,25 +330,30 @@ class CMRBackend(BaseBackend):
         def _reader(asset: Asset, x: int, y: int, z: int, **kwargs: Any) -> ImageData:
             s3_credentials = self._get_s3_credentials(asset)
 
-            if isinstance(self.reader, type) and self.reader == Reader:
-                aws_session = self._create_aws_session(s3_credentials)
+            if any(
+                field.name == "opener_options" for field in attr.fields(self.reader)
+            ):
+                options = self._build_reader_options(s3_credentials)
 
-                with rasterio.Env(aws_session):
-                    with self.reader(
+                with self.reader(
+                    asset["url"],
+                    tms=self.tms,  # type: ignore
+                    **options,
+                ) as src_dst:
+                    return src_dst.tile(x, y, z, **kwargs)
+            else:
+                with (
+                    rasterio.Env(
+                        self._create_aws_session(s3_credentials),
+                        **self.rasterio_env_kwargs,
+                    ),
+                    self.reader(
                         asset["url"],
-                        tms=self.tms,
+                        tms=self.tms,  # type: ignore
                         **self.reader_options,
-                    ) as src_dst:
-                        return src_dst.tile(x, y, z, **kwargs)
-
-            options = self._build_reader_options(s3_credentials)
-
-            with self.reader(
-                asset["url"],
-                tms=self.tms,
-                **options,
-            ) as src_dst:
-                return src_dst.tile(x, y, z, **kwargs)
+                    ) as src_dst,
+                ):
+                    return src_dst.tile(x, y, z, **kwargs)
 
         logger.info("reading assets")
         return mosaic_reader(mosaic_assets, _reader, tile_x, tile_y, tile_z, **kwargs)
@@ -335,9 +377,11 @@ class CMRBackend(BaseBackend):
         dst_crs: Optional[CRS] = None,
         bounds_crs: CRS = WGS84_CRS,
         bands_regex: Optional[str] = None,
+        access: Access | None = None,
         **kwargs: Any,
     ) -> Tuple[ImageData, List[str]]:
         """Create an Image from multiple items for a bbox."""
+        access = access or s3_auth_config.access
         xmin, ymin, xmax, ymax = bbox
 
         mosaic_assets = self.assets_for_bbox(
@@ -346,7 +390,7 @@ class CMRBackend(BaseBackend):
             xmax,
             ymax,
             coord_crs=bounds_crs,
-            access=s3_auth_config.access,
+            access=access,
             bands_regex=bands_regex,
             **cmr_query,
         )
@@ -357,23 +401,28 @@ class CMRBackend(BaseBackend):
         def _reader(asset: Asset, bbox: BBox, **kwargs: Any) -> ImageData:
             s3_credentials = self._get_s3_credentials(asset)
 
-            if isinstance(self.reader, type) and self.reader == Reader:
-                aws_session = self._create_aws_session(s3_credentials)
+            if any(
+                field.name == "opener_options" for field in attr.fields(self.reader)
+            ):
+                options = self._build_reader_options(s3_credentials)
 
-                with rasterio.Env(aws_session):
-                    with self.reader(
-                        asset["url"],
+                with self.reader(
+                    asset["url"],  # type: ignore
+                    **options,
+                ) as src_dst:
+                    return src_dst.part(bbox, **kwargs)
+            else:
+                with (
+                    rasterio.Env(
+                        self._create_aws_session(s3_credentials),
+                        **self.rasterio_env_kwargs,
+                    ),
+                    self.reader(
+                        asset["url"],  # type: ignore
                         **self.reader_options,
-                    ) as src_dst:
-                        return src_dst.part(bbox, **kwargs)
-
-            options = self._build_reader_options(s3_credentials)
-
-            with self.reader(
-                asset["url"],
-                **options,
-            ) as src_dst:
-                return src_dst.part(bbox, **kwargs)
+                    ) as src_dst,
+                ):
+                    return src_dst.part(bbox, **kwargs)
 
         return mosaic_reader(
             mosaic_assets,
@@ -391,9 +440,12 @@ class CMRBackend(BaseBackend):
         dst_crs: Optional[CRS] = None,
         shape_crs: CRS = WGS84_CRS,
         bands_regex: Optional[str] = None,
+        access: Access | None = None,
         **kwargs: Any,
     ) -> Tuple[ImageData, List[str]]:
         """Create an Image from multiple items for a GeoJSON feature."""
+        access = access or s3_auth_config.access
+
         if "geometry" in shape:
             shape = shape["geometry"]
 
@@ -406,7 +458,7 @@ class CMRBackend(BaseBackend):
 
         mosaic_assets = self.get_assets(
             *shape_bounds,
-            access=s3_auth_config.access,
+            access=access,
             bands_regex=bands_regex,
             **cmr_query,
         )
@@ -417,23 +469,28 @@ class CMRBackend(BaseBackend):
         def _reader(asset: Asset, shape: Dict, **kwargs: Any) -> ImageData:
             s3_credentials = self._get_s3_credentials(asset)
 
-            if isinstance(self.reader, type) and self.reader == Reader:
-                aws_session = self._create_aws_session(s3_credentials)
+            if any(
+                field.name == "opener_options" for field in attr.fields(self.reader)
+            ):
+                options = self._build_reader_options(s3_credentials)
 
-                with rasterio.Env(aws_session):
-                    with self.reader(
-                        asset["url"],
+                with self.reader(
+                    asset["url"],  # type: ignore
+                    **options,
+                ) as src_dst:
+                    return src_dst.feature(shape, **kwargs)
+            else:
+                with (
+                    rasterio.Env(
+                        self._create_aws_session(s3_credentials),
+                        **self.rasterio_env_kwargs,
+                    ),
+                    self.reader(
+                        asset["url"],  # type: ignore
                         **self.reader_options,
-                    ) as src_dst:
-                        return src_dst.feature(shape, **kwargs)
-
-            options = self._build_reader_options(s3_credentials)
-
-            with self.reader(
-                asset["url"],
-                **options,
-            ) as src_dst:
-                return src_dst.feature(shape, **kwargs)
+                    ) as src_dst,
+                ):
+                    return src_dst.feature(shape, **kwargs)
 
         return mosaic_reader(
             mosaic_assets,