+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add Earthdata Login based S3 access, except for the rasterio backend, to be
+  addressed separately.
+  ([#112](https://github.com/developmentseed/titiler-cmr/pull/112))
+
 ### Fixed
 
 - Fixed the missing parameter error message "`indexes` must be provided if not providing `bands_regex` and `bands`". The [alias](https://fastapi.tiangolo.com/tutorial/query-params-str-validations/#alias-parameters) for the `indexes` parameter is `bidx`, so the `bidx` parameter is what is missing from the parameter string. ([#101](https://github.com/developmentseed/titiler-cmr/pull/101))
@@ -93,7 +99,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Ability to run locally with Earthdata authentication ([#28](https://github.com/developmentseed/titiler-cmr/pull/28))
 
-[Unreleased]: <https://github.com/developmentseed/titiler-cmr/compare/v0.2.0..develop>
+[Unreleased]: <https://github.com/developmentseed/titiler-cmr/compare/v0.3.0..develop>
+[v0.3.0]: <https://github.com/developmentseed/titiler-cmr/compare/v0.2.0..v0.3.0>
 [v0.2.0]: <https://github.com/developmentseed/titiler-cmr/compare/v0.1.5..v0.2.0>
 [v0.1.5]: <https://github.com/developmentseed/titiler-cmr/compare/v0.1.4..v0.1.5>
 [v0.1.4]: <https://github.com/developmentseed/titiler-cmr/compare/v0.1.3..v0.1.4>